requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
exchangelib==5.4.3
icalendar==6.3.2
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
keyring==25.7.0
more-itertools==10.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
